+++ libdispatch/nclist.c
@@ -53,7 +53,7 @@ Free a list and its contents
 int
 nclistfreeall(NClist* l)
 {
-  int i;
+  unsigned long i;
   if(l == NULL) return TRUE;
   for(i=0;i<l->length;i++) {
       void* value = l->content[i];
@@ -110,11 +110,11 @@ nclistset(NClist* l, unsigned long index, void* elem)
 int
 nclistinsert(NClist* l, unsigned long index, void* elem)
 {
-  int i; /* do not make unsigned */
+  long i; /* do not make unsigned */
   if(l == NULL) return FALSE;
   if(index > l->length) return FALSE;
   nclistsetalloc(l,0);
-  for(i=(int)l->length;i>index;i--) l->content[i] = l->content[i-1];
+  for(i=(long)l->length;i>index;i--) l->content[i] = l->content[i-1];
   l->content[index] = elem;
   l->length++;
   return TRUE;
